Moved CLI_Rendering_Engine Library to external file
